requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.6
Django==5.2.3
django-crispy-forms==2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
inflection==0.5.1
narwhals==1.44.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
plotly==6.2.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
